+++ src/db.h
@@ -4,6 +4,7 @@ extern int		newmobs;
 extern int		newobjs;
 extern MOB_INDEX_DATA 	* mob_index_hash          [MAX_KEY_HASH];
 extern OBJ_INDEX_DATA 	* obj_index_hash          [MAX_KEY_HASH];
+extern ROOM_INDEX_DATA	* room_index_hash         [MAX_KEY_HASH];
 extern int		top_mob_index;
 extern int		top_obj_index;
 extern int  		top_affect;